refactor: reorganize task module and align tests with modular monolith structure
