+++ tests/integration/tasks/test_task_delete.py
