+++ tests/unit/tasks/application/test_task_service.py
@@ -1,0 +1,521 @@
+# tests/unit/services/test_task_service.py
+
+from datetime import date, datetime
+from unittest.mock import Mock
+
+import pytest
+
+from app.modules.tasks.domain.task import Task
+from app.modules.tasks.domain.task_status import TaskStatus
+from app.modules.tasks.domain.task_errors import InvalidTaskDueDateError
+from app.modules.tasks.application.task_repository import TaskRepository, TaskCreateData
+from app.modules.tasks.application.task_service import TaskNotFoundError, TaskService
+from tests.factories.task import build_task, future_date, future_date
+from app.modules.tasks.application.task_dtos import CreateTaskInput, UpdateTaskInput, PatchTaskInput
+
+valid_due_date = future_date()
+
+@pytest.fixture
+def repository() -> Mock:
+    """
+    Return a mock TaskRepository for each test.
+    """
+    return Mock(spec=TaskRepository)
+
+
+@pytest.fixture
+def task_service(repository: Mock) -> TaskService:
+    """
+    Return a TaskService instance with a mocked repository.
+    """
+    return TaskService(repository=repository)
+
+
+def save_task_side_effect(task: Task) -> Task:
+    return task
+
+
+# ----------------------------------------
+# get_task
+# ----------------------------------------
+
+def test_get_task_returns_task_when_repository_finds_it(task_service: TaskService, repository: Mock):
+    """
+    get_task should return the task when the repository finds it.
+    """
+    # Arrange
+    existing_task = build_task(task_id=1, title="Task 1")
+    repository.get_task.return_value = existing_task
+
+    # Act
+    result = task_service.get_task(1)
+
+    # Assert
+    assert result == existing_task
+    repository.get_task.assert_called_once_with(1)
+
+def test_get_task_raises_when_repository_returns_none(task_service: TaskService, repository: Mock):
+    """
+    get_task should raise TaskNotFoundError when the repository returns None.
+    """
+    # Arrange
+    repository.get_task.return_value = None
+
+    # Act / Assert
+    with pytest.raises(TaskNotFoundError, match="Task not found"):
+        task_service.get_task(999)
+
+    repository.get_task.assert_called_once_with(999)
+
+# ----------------------------------------
+# delete_task
+# ----------------------------------------
+
+def test_delete_task_calls_repository_delete_when_task_exists(task_service: TaskService, repository: Mock):
+    """
+    delete_task should call repository.delete_task when the task exists.
+    """
+    # Arrange
+    existing_task = build_task(task_id=1, title="Task to delete")
+    repository.get_task.return_value = existing_task
+
+    # Act
+    task_service.delete_task(1)
+
+    # Assert
+    repository.get_task.assert_called_once_with(1)
+    repository.delete_task.assert_called_once_with(1)
+
+def test_delete_task_raises_when_repository_returns_none(task_service: TaskService, repository: Mock):
+    """
+    delete_task should raise TaskNotFoundError when the repository does not find the task.
+    """
+    # Arrange
+    repository.get_task.return_value = None
+
+    # Act / Assert
+    with pytest.raises(TaskNotFoundError, match="Task not found"):
+        task_service.delete_task(999)
+
+    repository.get_task.assert_called_once_with(999)
+    repository.delete_task.assert_not_called()
+
+# ----------------------------------------
+# list_tasks
+# ----------------------------------------
+
+def test_list_tasks_returns_tasks_from_repository(task_service: TaskService, repository: Mock):
+    """
+    list_tasks should return the tasks provided by the repository.
+    """
+    # Arrange
+    stored_tasks = [build_task(task_id=1, title="Task 1"), build_task(task_id=2, title="Task 2")]
+    repository.list_tasks.return_value = stored_tasks
+
+    # Act
+    result = task_service.list_tasks()
+
+    # Assert
+    assert result == stored_tasks
+    repository.list_tasks.assert_called_once_with()
+
+# ----------------------------------------
+# create_task
+# ----------------------------------------
+
+def test_create_task_delegates_creation_to_repository_with_default_status(
+    task_service: TaskService,
+    repository: Mock,
+) -> None:
+    """
+    create_task should build TaskCreateData and delegate creation to the repository.
+    """
+    # Arrange
+    task_input = CreateTaskInput(
+        title="My first task",
+        description="testing",
+        due_date=None,
+        is_blocked=False,
+    )
+    created_task = build_task(
+        task_id=1,
+        title="My first task",
+        description="testing",
+        due_date=None,
+        is_blocked=False,
+    )
+    repository.create_task.return_value = created_task
+
+    # Act
+    result = task_service.create_task(task_input)
+
+    # Assert
+    assert result == created_task
+    repository.create_task.assert_called_once()
+
+    create_data = repository.create_task.call_args.args[0]
+    assert isinstance(create_data, TaskCreateData)
+    assert create_data.title == "My first task"
+    assert create_data.description == "testing"
+    assert create_data.status == TaskStatus.PENDING
+    assert create_data.due_date is None
+    assert create_data.is_blocked is False
+    assert isinstance(create_data.created_at, datetime)
+
+
+def test_create_task_passes_due_date_to_repository(
+    task_service: TaskService,
+    repository: Mock,
+) -> None:
+    """
+    create_task should pass due_date to the repository when provided.
+    """
+    # Arrange
+    task_input = CreateTaskInput(
+        title="Task with due date",
+        description="testing",
+        due_date=valid_due_date,
+        is_blocked=False,
+    )
+    created_task = build_task(
+        task_id=1,
+        title="Task with due date",
+        description="testing",
+        due_date=valid_due_date,
+        is_blocked=False,
+    )
+    repository.create_task.return_value = created_task
+
+    # Act
+    result = task_service.create_task(task_input)
+
+    # Assert
+    assert result == created_task
+    repository.create_task.assert_called_once()
+
+    create_data = repository.create_task.call_args.args[0]
+    assert isinstance(create_data, TaskCreateData)
+    assert create_data.title == "Task with due date"
+    assert create_data.description == "testing"
+    assert create_data.status == TaskStatus.PENDING
+    assert create_data.due_date == valid_due_date
+    assert create_data.is_blocked is False
+    assert isinstance(create_data.created_at, datetime)
+
+
+def test_create_task_passes_is_blocked_to_repository(
+    task_service: TaskService,
+    repository: Mock,
+) -> None:
+    """
+    create_task should pass is_blocked to the repository when provided.
+    """
+    # Arrange
+    task_input = CreateTaskInput(
+        title="Blocked task",
+        description="testing",
+        due_date=None,
+        is_blocked=True,
+    )
+    created_task = build_task(
+        task_id=1,
+        title="Blocked task",
+        description="testing",
+        due_date=None,
+        is_blocked=True,
+    )
+    repository.create_task.return_value = created_task
+
+    # Act
+    result = task_service.create_task(task_input)
+
+    # Assert
+    assert result == created_task
+    repository.create_task.assert_called_once()
+
+    create_data = repository.create_task.call_args.args[0]
+    assert isinstance(create_data, TaskCreateData)
+    assert create_data.title == "Blocked task"
+    assert create_data.description == "testing"
+    assert create_data.status == TaskStatus.PENDING
+    assert create_data.due_date is None
+    assert create_data.is_blocked is True
+    assert isinstance(create_data.created_at, datetime)
+
+# ----------------------------------------
+# update_task
+# ----------------------------------------
+
+def test_update_task_replaces_task_data_and_saves_it(task_service: TaskService, repository: Mock):
+    """
+    update_task should fully replace task fields and save the updated task.
+    """
+    # Arrange
+    existing_task = build_task(task_id=1, title="Original title", description="original", due_date=None)
+
+    repository.get_task.return_value = existing_task
+
+    task_update = UpdateTaskInput(
+        title="Updated title",
+        description="updated",
+        status=TaskStatus.IN_PROGRESS,
+        due_date=valid_due_date,
+        is_blocked=True
+    )
+
+    repository.save_task.side_effect = save_task_side_effect
+
+    # Act
+    result = task_service.update_task(1, task_update)
+
+    # Assert
+    assert result.id == 1
+    assert result.title == "Updated title"
+    assert result.description == "updated"
+    assert result.status == TaskStatus.IN_PROGRESS
+    assert result.due_date == valid_due_date
+    assert result.is_blocked is True
+    assert result.created_at == existing_task.created_at
+    assert result.updated_at is not None
+
+    repository.get_task.assert_called_once_with(1)
+    repository.save_task.assert_called_once_with(existing_task)
+
+def test_update_task_raises_when_repository_returns_none(task_service: TaskService, repository: Mock):
+    """
+    update_task should raise TaskNotFoundError when the repository does not find the task.
+    """
+    # Arrange
+    repository.get_task.return_value = None
+    task_update = UpdateTaskInput(title="Updated title", description="updated", status=TaskStatus.IN_PROGRESS, due_date=date(2026, 3, 20), is_blocked=False)
+
+    # Act / Assert
+    with pytest.raises(TaskNotFoundError, match="Task not found"):
+        task_service.update_task(999, task_update)
+
+    repository.get_task.assert_called_once_with(999)
+    repository.save_task.assert_not_called()
+
+def test_update_task_saves_pending_state_with_nullable_fields(task_service: TaskService, repository: Mock):
+    """
+    update_task should allow description and due_date to be set to None.
+    """
+    # Arrange
+    existing_task = build_task(task_id=1, title="Original title", description="original")
+    repository.get_task.return_value = existing_task
+    repository.save_task.side_effect = save_task_side_effect
+
+    task_update = UpdateTaskInput(
+        title="Updated title",
+        description=None,
+        status=TaskStatus.PENDING,
+        due_date=None,
+        is_blocked=False
+    )
+
+    # Act
+    result = task_service.update_task(1, task_update)
+
+    # Assert
+    assert result.id == 1
+    assert result.title == "Updated title"
+    assert result.description is None
+    assert result.status == TaskStatus.PENDING
+    assert result.due_date is None
+    assert result.is_blocked is False
+    assert result.updated_at is not None
+
+    repository.get_task.assert_called_once_with(1)
+    repository.save_task.assert_called_once_with(existing_task)
+
+def test_update_task_does_not_save_when_no_real_changes(task_service: TaskService, repository: Mock):
+    """
+    update_task should not save the task when the payload does not change any field.
+    """
+    # Arrange
+    existing_task = build_task(
+        task_id=1,
+        title="Original title",
+        description="original",
+        status=TaskStatus.IN_PROGRESS,
+        due_date=valid_due_date
+    )
+    repository.get_task.return_value = existing_task
+
+    task_update = UpdateTaskInput(
+        title="Original title",
+        description="original",
+        status=TaskStatus.IN_PROGRESS,
+        due_date=valid_due_date,
+        is_blocked=False
+    )
+
+    # Act
+    result = task_service.update_task(1, task_update)
+
+    # Assert
+    assert result is existing_task
+    assert result.updated_at is None
+
+    repository.get_task.assert_called_once_with(1)
+    repository.save_task.assert_not_called()
+
+def test_update_task_updates_is_blocked_when_provided(task_service: TaskService, repository: Mock):
+    """
+    update_task should replace is_blocked when provided in the PUT payload.
+    """
+    # Arrange
+    existing_task = build_task(task_id=1, status=TaskStatus.PENDING, due_date=None)
+    repository.get_task.return_value = existing_task
+    repository.save_task.side_effect = save_task_side_effect
+
+    task_update = UpdateTaskInput(
+        title=existing_task.title,
+        description=existing_task.description,
+        status=TaskStatus.PENDING,
+        due_date=None,
+        is_blocked=True
+    )
+
+    # Act
+    result = task_service.update_task(1, task_update)
+
+    # Assert
+    assert result.is_blocked is True
+    assert result.updated_at is not None
+    repository.get_task.assert_called_once_with(1)
+    repository.save_task.assert_called_once_with(existing_task)
+
+def test_update_task_propagates_domain_error_and_does_not_save_when_final_state_is_invalid(task_service: TaskService, repository: Mock) -> None:
+    existing_task = build_task(task_id=1, status=TaskStatus.PENDING, due_date=None)
+    repository.get_task.return_value = existing_task
+
+    task_update = UpdateTaskInput(
+        title=existing_task.title,
+        description=existing_task.description,
+        status=TaskStatus.IN_PROGRESS,
+        due_date=None,
+        is_blocked=False
+    )
+
+    with pytest.raises(InvalidTaskDueDateError, match="required for IN_PROGRESS tasks"):
+        task_service.update_task(1, task_update)
+
+    repository.get_task.assert_called_once_with(1)
+    repository.save_task.assert_not_called()
+
+# ----------------------------------------
+# patch_task
+# ----------------------------------------
+
+def test_patch_task_updates_only_provided_fields(task_service: TaskService, repository: Mock) -> None:
+    existing_task = build_task(
+        task_id=1,
+        title="Original title",
+        description="original",
+        status=TaskStatus.PENDING,
+        due_date=valid_due_date,
+        is_blocked=False,
+    )
+    repository.get_task.return_value = existing_task
+    repository.save_task.side_effect = save_task_side_effect
+
+    task_patch = PatchTaskInput(
+        description="updated description",
+        description_provided=True,
+    )
+
+    result = task_service.patch_task(1, task_patch)
+
+    assert result.title == "Original title"
+    assert result.description == "updated description"
+    assert result.status == TaskStatus.PENDING
+    assert result.due_date == valid_due_date
+    assert result.is_blocked is False
+    assert result.updated_at is not None
+
+    repository.get_task.assert_called_once_with(1)
+    repository.save_task.assert_called_once_with(existing_task)
+
+def test_patch_task_allows_description_to_be_cleared(task_service: TaskService, repository: Mock) -> None:
+    existing_task = build_task(
+        task_id=1,
+        description="original",
+        due_date=valid_due_date,
+    )
+    repository.get_task.return_value = existing_task
+    repository.save_task.side_effect = save_task_side_effect
+
+    task_patch = PatchTaskInput(
+        description=None,
+        description_provided=True,
+    )
+
+    result = task_service.patch_task(1, task_patch)
+
+    assert result.description is None
+    assert result.updated_at is not None
+
+    repository.get_task.assert_called_once_with(1)
+    repository.save_task.assert_called_once_with(existing_task)
+
+def test_patch_task_does_not_save_when_patch_does_not_change_state(task_service: TaskService, repository: Mock) -> None:
+    existing_task = build_task(
+        task_id=1,
+        title="Original title",
+        description="original",
+        status=TaskStatus.PENDING,
+        due_date=valid_due_date,
+        is_blocked=False,
+    )
+    repository.get_task.return_value = existing_task
+
+    task_patch = PatchTaskInput(
+        title="Original title",
+        title_provided=True,
+    )
+
+    result = task_service.patch_task(1, task_patch)
+
+    assert result is existing_task
+    assert result.updated_at is None
+
+    repository.get_task.assert_called_once_with(1)
+    repository.save_task.assert_not_called()
+
+def test_patch_task_raises_when_repository_returns_none(task_service: TaskService, repository: Mock) -> None:
+    repository.get_task.return_value = None
+
+    task_patch = PatchTaskInput(
+        description="updated",
+        description_provided=True,
+    )
+
+    with pytest.raises(TaskNotFoundError, match="Task not found"):
+        task_service.patch_task(999, task_patch)
+
+    repository.get_task.assert_called_once_with(999)
+    repository.save_task.assert_not_called()
+
+def test_patch_task_propagates_domain_error_and_does_not_save_when_final_state_is_invalid(
+    task_service: TaskService,
+    repository: Mock,
+) -> None:
+    existing_task = build_task(
+        task_id=1,
+        status=TaskStatus.PENDING,
+        due_date=None,
+        is_blocked=False,
+    )
+    repository.get_task.return_value = existing_task
+
+    task_patch = PatchTaskInput(
+        status=TaskStatus.IN_PROGRESS,
+        status_provided=True,
+    )
+
+    with pytest.raises(InvalidTaskDueDateError, match="required for IN_PROGRESS tasks"):
+        task_service.patch_task(1, task_patch)
+
+    repository.get_task.assert_called_once_with(1)
+    repository.save_task.assert_not_called()
